refactor(schema): every migration comes from a model, and the models say more
The corpus had one hand-written migration in it, to drop the foreign keys the
cascade ones replaced. Needing it was the signal that the models were wrong
rather than the generator: buddy migrate:regenerate rebuilt the whole corpus
from them - 107 files from 104 models, replacing 139 - which puts every
constraint inline on its CREATE TABLE and leaves the repair nothing to do.
There is no real data yet, so replaying from scratch cost nothing.

Regenerating surfaced nine foreign keys that existed only as attributes. Two
came from the framework's own audit; reading the emitted DDL for _id columns
with no REFERENCES found seven more, all of them a role rather than an owner:
closed_by_id, merged_by_id, stack_parent_id, resolved_by_id, revoked_by_id,
requested_by_id and repositories.parent_id. None had a constraint, so a row
could point at a user, milestone or repository that no longer existed.

They are declared now with the action each actually wants. cascade where the
row is meaningless without its parent; set null everywhere else, because an
issue outlives whoever closed it and a thread stays resolved when that account
is gone. repositories.parent_id is the one worth calling out: a fork now
detaches from a deleted upstream rather than going with it, which is what
purge.ts does by hand today.

tests/unit/migrations-from-models.test.ts keeps it this way. It refuses a file
the generator would not have named, refuses hand-written commentary in the
corpus - explain it in the model, where it survives a regeneration - and
checks that every relation declaring an onDelete reaches a real constraint. A
model that says the database will clean up while the database does nothing is
worse than saying nothing at all.

+++ database/migrations/0000000013-create-labels-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS "labels" (
   "id" BIGSERIAL PRIMARY KEY,
-  "board_id" integer not null REFERENCES "boards"("id"),
+  "board_id" bigint not null REFERENCES "boards"("id"),
   "name" varchar(60) not null,
   "color" varchar(40),
   "created_at" timestamp not null default CURRENT_TIMESTAMP,
