+++ database/migrations/0000000044-create-access_token_permissions-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS "access_token_permissions" (
   "id" BIGSERIAL PRIMARY KEY,
-  "access_token_id" integer not null REFERENCES "access_tokens"("id"),
+  "access_token_id" bigint not null REFERENCES "access_tokens"("id"),
   "scope" "access_token_permissions_scope_type",
   "level" "access_token_permissions_level_type" default 'read',
   "created_at" timestamp not null default CURRENT_TIMESTAMP,
